# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Робоскоп API",
    "version": "1.0.0",
    "description": "Сессионная авторизация Django. Для POST/PUT/DELETE передавайте cookie csrftoken как заголовок X-CSRFToken. GET / устанавливает CSRF cookie. Все суммы в рублях. Неизвестные входные поля отклоняются. Экспорт сохранённой ревизии использует её снимок. При сохранении ранее загруженной версии calculation_revision указывает снимок для воспроизведения, а revision задаёт текущую ревизию проекта для защиты от конфликтов."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/auth/{action}/": {
      "post": {
        "summary": "Вход, регистрация или выход",
        "responses": {
          "200": {
            "description": "Успешный результат"
          },
          "400": {
            "description": "Ошибка входов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Требуется вход"
          },
          "403": {
            "description": "Нет доступа или неверный CSRF"
          },
          "404": {
            "description": "Запись не найдена или принадлежит другому пользователю"
          },
          "409": {
            "description": "Конфликт ревизии или активный прогон"
          },
          "429": {
            "description": "Превышен лимит"
          }
        },
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "login",
                "logout"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/catalog/import/": {
      "post": {
        "summary": "Транзакционное обновление CSV-каталога администратором",
        "responses": {
          "200": {
            "description": "Успешный результат"
          },
          "400": {
            "description": "Ошибка входов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Требуется вход"
          },
          "403": {
            "description": "Нет доступа или неверный CSRF"
          },
          "404": {
            "description": "Запись не найдена или принадлежит другому пользователю"
          },
          "409": {
            "description": "Конфликт ревизии или активный прогон"
          },
          "429": {
            "description": "Превышен лимит"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRFToken"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "fields": {
            "type": "object"
          }
        }
      }
    }
  }
}